requirements.txt
requests>=2.28.1
qrcode>=7.3.1
rich>=12.6.0
pycryptodome>=3.15.0
lxml>=4.9.1
pyyaml>=6.0
jsonpath-python>=1.0.6
beautifulsoup4>=4.11.1
ddddocr>=1.4.10
numpy>=1.24.3
opencv-python>=4.7.0.72
dataclasses-json>=0.5.7
yarl>=1.9.2
openai>=1.14.3
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0